toolkit/components/passwordmgr/test/test_input_events_for_identical_values.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for input events in Login Manager when username/password are filled in already</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="onNewEvent(event)">
Login Manager test: input events should fire.

<script>
commonInit();
SimpleTest.waitForExplicitFinish();

/** Test for Login Manager: form fill when form is already filled, should not get input events. **/

var onloadFired = false;

function onNewEvent(e) {
  console.error("Got " + e.type + " event.");
  if (e.type == "load") {
    onloadFired = true;
    $_(1, "uname").focus();
    sendKey("Tab");
  } else {
    ok(false, "Got an input event for " + e.target.name + " field, which shouldn't happen.");
  }
}

SimpleTest.registerCleanupFunction(function cleanup() {
  $_(1, "uname").removeAttribute("oninput");
  $_(1, "pword").removeAttribute("onfocus");
  $_(1, "pword").removeAttribute("oninput");
  document.body.removeAttribute("onload");
});
</script>

<p id="display"></p>

<div id="content">

  <form id="form1" action="formtest.js">
    <p>This is form 1.</p>
    <input  type="text"       name="uname" oninput="onNewEvent(event)" value="testuser">
    <input  type="password"   name="pword" oninput="onNewEvent(event)" onfocus="setTimeout(function() { SimpleTest.finish() }, 1000);" value="testpass">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

</div>
<pre id="test"></pre>
</body>
</html>
